+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - New #104: Add methods `StringHelper::trim()`, `StringHelper::ltrim()`, `StringHelper::rtrim()` (@olegbaturin)
 - Enh #83: Make minor refactoring with Rector help (@vjik)
 - Enh #111: Minor refactoring (@Tigrov)
+- Enh #92: Add `$strict` parameter to `Inflector::toSnakeCase()` method (@arogachev)
 
 ## 2.1.2 July 27, 2023